bsd: add syscall7, fix includes and types, improve portability
- Add __syscall7 for BSD on multiple architectures
  for 7-argument syscall support.
- Add and fix missing includes in BSD headers
  and sources for portability, types, and build correctness.
- Adjust struct fields, typedefs, enums, and macros
  for ABI compatibility and clearer BSD support.
- Fix and improve logic in several BSD sources
  (sethostname, ktrace, minherit, bsdlabel, readlabel, etc).
- Use _BSD_SOURCE where needed for visibility of BSD features.
- Improve mapping of constants, add missing defines,
  and modernize enum and macro usage.
- General polish for multi-OS support and easier maintenance
  between BSDs, and Linux targets.

+++ src/bsd/sbrk.c
@@ -7,7 +7,7 @@
 extern char _end;
 static void *__curbrk = &_end;
 
-void *sbrk(int inc)
+void *sbrk(intptr_t inc)
 {
 	void *oldbrk = __curbrk;
